Rename tgan_synth -> model
