+++ tgan/model.py
